Improve FirebaseAuthUI cancellation handling
This commit refines the logic for handling user cancellation of the FirebaseAuthUI flow.

Previously, the `FauthSignInResult.Destroy` event was triggered even if the app went to the background during the authentication process and then the user cancelled.

Now, `FauthSignInResult.Destroy` is only emitted if the cancellation occurs while the app is in the foreground. This is achieved by tracking the app's background state using `ScreenManager.screenState` and the new `screenNameOfFirebaseAuthUiLauncher` property.

Additionally, the library version has been bumped to 1.1.6.