Bump firebase-ui-auth to 9.1.1-SNAPSHOT and library version to 1.3.0
This commit updates the `firebase-ui-auth` dependency to version `9.1.1-SNAPSHOT` and bumps the library version to `1.3.0`.

Additionally, it introduces a fix to hide the email sign-up option when `allowNewAccounts` is set to `false` in the email provider configuration.

The `mavenLocal()` repository has also been added to the plugin and dependency management sections in `settings.gradle.kts`.